backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
python-dotenv
passlib
fastapi-mail
python-jose[cryptography]
psycopg2
